requirements.txt
pyarrow==14.0.0
langchain==0.3.8
langchain_community==0.3.8
langchain_core==0.3.21
langchain_openai==0.2.9
python-dotenv==1.0.1
streamlit==1.40.1
faiss-cpu==1.9.0.post1